Add Type.array()
