+++ isly.ts
@@ -1,5 +1,4 @@
 export { any } from "./any"
-export { array } from "./array"
 export { boolean } from "./boolean"
 export { Flaw } from "./Flaw"
 export { fromIs } from "./fromIs"
@@ -12,5 +11,5 @@ export { object } from "./object"
 export { record } from "./record"
 export { string } from "./string"
 export { tuple } from "./tuple"
-export { Type } from "./Type"
+export { Type, array } from "./Type"
 export { union } from "./union"
